renaming ecg to beat
