+++ src/vg_beat_detectors/__init__.py
@@ -1,0 +1,3 @@
+from vg_beat_detectors.vg_detector import VisGraphDetector
+from vg_beat_detectors.fast_whvg import FastWHVG
+from vg_beat_detectors.fast_nvg import FastNVG
