+++ src/vg_beat_detectors/vg_detector.py
